requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
geographiclib==1.52
geopy==2.2.0
googlemaps==4.6.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
populartimes @ file:///home/sargis/PycharmProjects/MapsGoogle/populartimes
python-dotenv==0.20.0
requests==2.28.0
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
